Invest_iq_APP/requirements.txt
requests
python-dotenv
tabulate
Flask
flask_caching
redis